pyproject.toml
[tool.poetry]
name = "sample"
version = "0.1.0"
description = ""
authors = ["hemanta212 <sharmahemanta.212@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
psycopg2 = "^2.9.6"
playwright = "^1.34.0"
free-proxy = "^1.1.1"
pytz = "^2023.3"
loguru = "^0.7.0"
bs4 = "^0.0.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ptpython = "^3.0.23"
rich = "^13.4.1"
black = "^23.3.0"
rope = "^1.8.0"
pylint = "^2.17.4"
pylsp-mypy = "^0.6.7"
pyls-isort = "^0.2.2"
pylsp-rope = "^0.1.11"
autoflake = "^2.1.1"
radon = "^6.0.1"
lxml-stubs = "^0.4.0"
types-beautifulsoup4 = "^4.12.0.5"
types-psycopg2 = "^2.9.21.10"
types-pytz = "^2023.3.0.0"
types-requests = "^2.31.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
# make it compatible with black
profile = "black" 